+++ app/lab/murajaah-at-taisir/components/MenuButton.tsx
@@ -58,7 +58,7 @@ export function MenuButton({ className }: { className?: string }) {
             className="bg-white fixed bottom-0 left-0 right-0 rounded-t-xl pb-8 pt-5 slide-up-active"
           >
             <div className="mx-auto max-w-105 px-3 sm:px-5">
-              <h2 className="text-lg tracking-tight font-bold mb-3">
+              <h2 className="text-lg tracking-tight font-bold mb-4">
                 Pengaturan
               </h2>
               <div className="flex flex-col gap-5 max-h-[60vh] overflow-auto -mr-5 pr-5">